refactor: separate themeContext and themes in diff files
