+++ src/constants/themeContext.js
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+const initialContext = {
+  theme: 'light',
+  toggleTheme: () => {},
+};
+const ThemeContext = createContext(initialContext);
+export default ThemeContext;
